chore: Trying Augment to break down LdapUtils responsibilities
